% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMuPETFlow.R
\name{runMuPETFlow}
\alias{runMuPETFlow}
\title{Run the MuPET-Flow app}
\usage{
runMuPETFlow()
}
\description{
This function launches the Shiny app included in this package.
The first time the tool launches, it will try to install additional \strong{Bioconductor} dependencies that are not managed via \code{install.packages()}. This process might take a few minutes.
Once the application is launched, you can either:
\enumerate{
\item Load your experimental data.
\item Run an in-app example by clicking the 'Example' button.
}\if{html}{\out{
<br>
<br>For the first case, selecting the channel where the data was acquired is mandatory. For the second case, the example channel `FL4-A` is automatically detected for demonstration purposes.
}}
}
\details{
After launching the app, you can follow the app flow, which is divided into three tabs: \emph{\strong{Peaks}}, \emph{\strong{Regression}} and \emph{\strong{Summary}}.  Below is a general description of the options available in each tab:
\subsection{Peaks}{
\itemize{
\item \strong{Select a sample (optional):} Allows visual exploration of individual samples if desired.
\item \strong{Adjust smoothing (optional):} Adjusts the histogram curve for noisy samples.
\item \strong{Adjust window width (optional):} Defines the interval where the app will look for peaks.
\item \strong{Select minimum cell count to call a peak (optional):} Useful for samples with a low number of events.
\item \strong{Select maximum number of peaks to plot (optional):} Useful for samples with heterogeneous populations where more peaks are present.
}
}

\subsection{Regression}{
\itemize{
\item \strong{Select type of analysis:} Choose between "Ploidy" or "Genome size" analysis.
\item \strong{Select number of standards:} A minimum of two standards is required.
\item \strong{Select standard samples and values:} This is the ploidy or genome size of your standards.
}
}

\subsection{Summary}{
\itemize{
\item \strong{Results preview:} Creates a compiled figure with histograms for all samples.
\item \strong{Save plot:} Saves the histograms in either PNG or TIFF format with customizable size and quality. Optionally, you can control the grid layout.
\item \strong{Save table:} Exports the parameters used and the estimated ploidy or genome size as a CSV file.
}\if{html}{\out{
<br>
If any errors are detected, you can go back to the Peaks tab to review the parameters. Note that the regression must also be re-done after parameter adjustments.
}}

}
}
